feat(cli): call version check in another thread
